+++ src/timeseriesflattener/processors/static.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .._intermediary_frames import ProcessedFrame
+from ..intermediary import ProcessedFrame
 
 if TYPE_CHECKING:
     from ..feature_specs.prediction_times import PredictionTimeFrame
