+++ src/timeseriesflattener/processors/static_test.py
