+++ src/timeseriesflattener/processors/timedelta.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 import polars as pl
 
-from .._intermediary_frames import ProcessedFrame
+from ..intermediary import ProcessedFrame
 
 if TYPE_CHECKING:
     from ..feature_specs.prediction_times import PredictionTimeFrame
