+++ src/timeseriesflattener/specs/test_from_legacy.py
