+++ src/timeseriesflattener/specs/timedelta.py
@@ -3,13 +3,13 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Literal
 
-from .._frame_validator import _validate_col_name_columns_exist
-from .meta import ValueFrame
+from ..validators
+from .value import ValueFrame
 
 if TYPE_CHECKING:
     import polars as pl
 
-    from .timestamp_frame import TimestampValueFrame
+    from .timestamp import TimestampValueFrame
 
 
 @dataclass
@@ -31,7 +31,7 @@ class TimeDeltaSpec:
         """
 
     def __post_init__(self):
-        _validate_col_name_columns_exist(obj=self)
+        validate_col_name_columns_exist(obj=self)
         max_values_per_id = (
             self.init_frame.collect()
             .get_column(self.init_frame.entity_id_col_name)
