+++ src/timeseriesflattener/specs/value.py
@@ -7,8 +7,8 @@ from typing import Literal
 import pandas as pd
 import polars as pl
 
-from .._frame_validator import _validate_col_name_columns_exist
-from ..frame_utilities.anyframe_to_lazyframe import _anyframe_to_lazyframe
+from ..validators import validate_col_name_columns_exist
+from ..utils import anyframe_to_lazyframe
 
 
 @dataclass
@@ -30,11 +30,11 @@ class ValueFrame:
         self, init_df: pl.LazyFrame | pl.DataFrame | pd.DataFrame, coerce_to_lazy: bool
     ):
         if coerce_to_lazy:
-            self.df = _anyframe_to_lazyframe(init_df)
+            self.df = anyframe_to_lazyframe(init_df)
         else:
-            self.df: pl.LazyFrame = init_df
+            self.df: pl.LazyFrame = init_df  # type: ignore
 
-        _validate_col_name_columns_exist(obj=self)
+        validate_col_name_columns_exist(obj=self)
         self.value_col_names = [
             col
             for col in self.df.columns
@@ -59,7 +59,7 @@ class LookPeriod:
             )
 
 
-def _lookdistance_to_normalised_lookperiod(
+def lookdistance_to_normalised_lookperiod(
     lookdistance: dt.timedelta | tuple[dt.timedelta, dt.timedelta],
     direction: Literal["ahead", "behind"],
 ) -> LookPeriod:
